enhance conda post-link and pre-unlink scripts
Explicitly Pass "--prefix", the default should be fine but it is safer to be
explicit.

Simplify the scripts by removing the block, they are un-needed for single line
commands.

In the windows scripts set the PATH explicitly as most versions of conda do not
set this prior to calling the script. See:
https://github.com/conda/conda/issues/7789